+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # CHANGELOG
 
+- Allow ReRoute strategy in combination with DCM
 - Add TourLengthFilter and document filters
 - BC: Remove over-complicated generics for UtilitySelector/Factory
 - Update MATSim version, fix MainModeIdentifier and remove StageActivityTypes